Initial commit
It took a while to consolidate all of the backups and archives into one (mostly) hash-singular repo. This should be nearly all of the writing I've done over the years. I tried, each year, to turn my crazy ideas into a book. This resulted in the exploration of my core themes from a score of different angles. Its still in a very raw state, so read at your own peril!